+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 3.2.1
 * Optimized `TADPole` for multiple `k` and `dc` values.
 * Partitional and hierarchical configurations in `compare_clusterings` now take into account the `symmetric` control parameter if provided.
+* Optimized PAM centroids with `pam.precompute = FALSE` by using sparse matrices from the `Matrix` package.
 
 ## Version 3.2.0
 * Added functions `compare_clusterings` and helpers to compare many clustering configurations, possibly in parallel.